Code/project-6-stateful-scaling-prithiraj-bhuyan/p3django/pyproject.toml
[tool.poetry]
name = "p3django"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.3"
djangorestframework-simplejwt = {extras = ["crypto"], version = "^5.5.1"}
python-jwt = "^4.1.0"
djangorestframework = "^3.17.1"
dotenv = "^0.9.9"
pyjwt = "^2.12.1"
mysqlclient = "^2.2.8"

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
